mythx-bot/requirements.txt
xai-sdk>=0.1.0
requests>=2.32.0
requests-oauthlib>=1.3.1
python-dotenv>=1.0.0
supabase>=2.5.0
moviepy==1.0.3
